+++ gc2d/controller/button/choose_palette_button.py
