+++ gc2d/view/main_window.py
@@ -3,10 +3,10 @@ from PyQt5.QtWidgets import QMainWindow, QLabel
 from pyqtgraph import QtCore
 from pyqtgraph.dockarea import Dock, DockArea
 
-from gc2d.controller.choose_palette_button import ChoosePaletteButton
-from gc2d.controller.exit_button import ExitButton
-from gc2d.controller.open_button import OpenButton
-from gc2d.controller.draw_button import DrawButton
+from controller.button.choose_palette_button import ChoosePaletteButton
+from controller.button.exit_button import ExitButton
+from controller.button.open_button import OpenButton
+from controller.button.draw_button import DrawButton
 from gc2d.view.plot_2d_widget import Plot2DWidget
 from gc2d.view.plot_3d_widget import Plot3DWidget
 
@@ -86,8 +86,6 @@ class Window(QMainWindow):
 
         self.plot_2d = Plot2DWidget(self.model_wrapper, dock_2d)
         dock_2d.addWidget(self.plot_2d)
-       
-
     
     def keyPressEvent(self, event): 
         if event.key() == QtCore.Qt.Key_Escape:
